@@ -1,33 +1,37 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.6.3'
+ruby '3.0.0'
 
-gem 'rails', '~> 6.0.2', '>= 6.0.2.1'
-gem 'pg', '~> 1.2.1'
-gem 'puma', '~> 4.1'
-gem 'bootsnap', '>= 1.4.2', require: false
+gem 'bootsnap', '~> 1.18', '>= 1.18.4', require: false
+gem 'pg', '~> 1.5.8'
+gem 'puma', '~> 6.4', '>= 6.4.2'
+gem 'rails', '~> 7.1', '>= 7.1.3'
 
 group :development, :test do
-  gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
-  gem 'rspec-rails', '~> 3.9'
-  gem 'ffaker', '~> 2.13.0'
-  gem 'factory_bot_rails', '~> 5.1.1'
-  gem 'pry', '~> 0.12.2'
-  gem 'bullet', '~> 6.1.0'
-  gem 'i18n', '~> 1.8', '>= 1.8.1'
-
+  gem 'bullet', '~> 7.2'
+  gem 'byebug', platforms: %i[mri mingw x64_mingw]
+  gem 'factory_bot_rails', '~> 6.4', '>= 6.4.3'
+  gem 'ffaker', '~> 2.23'
+  gem 'i18n', '~> 1.14', '>= 1.14.5'
+  gem 'pry', '~> 0.14.2'
+  gem 'rspec-rails', '~> 7.0', '>= 7.0.1'
+  gem 'rubocop', '~> 1.66', '>= 1.66.1'
+  gem 'rubocop-factory_bot', '~> 2.26', '>= 2.26.1'
+  gem 'rubocop-rails', '~> 2.26', '>= 2.26.2'
+  gem 'rubocop-rspec', '~> 3.0', '>= 3.0.5'
+  gem 'rubocop-rspec_rails', '~> 2.30'
 end
 
 group :test do
-   gem 'simplecov', '~> 0.17.1', require: false
-   gem 'shoulda-matchers', '~> 4.2.0'
- end
-
-group :development do
-  gem 'listen', '>= 3.0.5', '< 3.2'
-  gem 'spring', '~> 2.1.0'
-  gem 'spring-watcher-listen', '~> 2.0.0'
+  gem 'shoulda-matchers', '~> 5.3'
+  gem 'simplecov', '~> 0.22.0'
 end
 
-gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
+group :development do
+  gem 'listen', '~> 3.9'
+  gem 'spring', '~> 4.2', '>= 4.2.1'
+  gem 'spring-watcher-listen', '~> 2.1'
+end
+
+gem 'tzinfo-data', platforms: %i[mingw mswin x64_mingw jruby]
