@@ -1,5 +1,5 @@
 class Inventory < ApplicationRecord
-  has_one :person
+  has_one :person, dependent: :destroy
 
   validates :water,       presence: true, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
   validates :food,        presence: true, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
