@@ -1,13 +1,14 @@
 class PercentageNonInfected
   def percentage
-    @result ||= (healthy_count/total_count.to_f) * 100
+    @result ||= (healthy_count / total_count.to_f) * 100
 
     return 0 if @result.nan?
+
     @result.round(2)
   end
 
   private
-  
+
   def healthy_count
     @healthy_count ||= Person.where(infected: false).count
   end
