@@ -1,5 +1,5 @@
 class AddInfectedToPeople < ActiveRecord::Migration[6.0]
   def change
-    add_column :people, :infected, :boolean
+    add_column :people, :infected, :boolean, default: false, null: false
   end
 end
