@@ -1,5 +1,5 @@
 class AddInventoryToPeople < ActiveRecord::Migration[6.0]
   def change
-    add_reference :people, :inventory, null: false, foreign_key: true
+    add_reference :people, :inventory, foreign_key: true
   end
 end
