@@ -2,16 +2,15 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_01_05_021747) do
-
+ActiveRecord::Schema[7.1].define(version: 2020_01_05_021747) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -20,8 +19,8 @@ ActiveRecord::Schema.define(version: 2020_01_05_021747) do
     t.integer "food"
     t.integer "medication"
     t.integer "ammunition"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "people", force: :cascade do |t|
@@ -29,8 +28,8 @@ ActiveRecord::Schema.define(version: 2020_01_05_021747) do
     t.integer "age"
     t.string "gender"
     t.string "local"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.boolean "infected"
     t.bigint "inventory_id", null: false
     t.index ["inventory_id"], name: "index_people_on_inventory_id"
@@ -39,8 +38,8 @@ ActiveRecord::Schema.define(version: 2020_01_05_021747) do
   create_table "report_people", force: :cascade do |t|
     t.bigint "report_to_id", null: false
     t.bigint "report_from_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["report_from_id"], name: "index_report_people_on_report_from_id"
     t.index ["report_to_id"], name: "index_report_people_on_report_to_id"
   end
