@@ -4,7 +4,7 @@ FactoryBot.define do
     age      { FFaker::Random.rand(1..99) }
     gender   { FFaker::GenderBR.maybe }
     infected { false }
-    local    { FFaker::Geolocation.lat.to_s + ", " + FFaker::Geolocation.lng.to_s }
+    local    { "#{FFaker::Geolocation.lat}, #{FFaker::Geolocation.lng}" }
   end
 
   factory :person_infected, class: 'Person' do
@@ -12,25 +12,24 @@ FactoryBot.define do
     age      { FFaker::Random.rand(1..99) }
     gender   { FFaker::GenderBR.maybe }
     infected { true }
-    local    { FFaker::Geolocation.lat.to_s + ", " + FFaker::Geolocation.lng.to_s }
+    local    { "#{FFaker::Geolocation.lat}, #{FFaker::Geolocation.lng}" }
   end
 
-  factory :person_with_inventory, class: 'Person'do
+  factory :person_with_inventory, class: 'Person' do
     name         { FFaker::NameBR.name }
     age          { FFaker::Random.rand(1..99) }
     gender       { FFaker::GenderBR.maybe }
     infected     { false }
-    local        { FFaker::Geolocation.lat.to_s + ", " + FFaker::Geolocation.lng.to_s }
+    local        { "#{FFaker::Geolocation.lat}, #{FFaker::Geolocation.lng}" }
     inventory_id { FactoryBot.create(:inventory_trade).id }
   end
 
-  factory :person_infected_with_inventory, class: 'Person'do
+  factory :person_infected_with_inventory, class: 'Person' do
     name         { FFaker::NameBR.name }
     age          { FFaker::Random.rand(1..99) }
     gender       { FFaker::GenderBR.maybe }
     infected     { true }
-    local        { FFaker::Geolocation.lat.to_s + ", " + FFaker::Geolocation.lng.to_s }
+    local        { "#{FFaker::Geolocation.lat}, #{FFaker::Geolocation.lng}" }
     inventory_id { FactoryBot.create(:inventory_trade).id }
   end
-  
 end
