@@ -1,44 +1,44 @@
-require "rails_helper"
+require 'rails_helper'
 
-RSpec.describe ReportPerson, type: :model do
-  describe 'validations' do
-    it { should belong_to(:report_to) } 
-    it { should belong_to(:report_from) }
-  end
-
-  let(:person_to)   { create(:person_with_inventory) }
+RSpec.describe ReportPerson do
   let(:person_from) { create(:person_with_inventory) }
+  let(:person_to)   { create(:person_with_inventory) }
 
-  context "Valid Report" do
-    it "It is a valid report" do
-      report = ReportPerson.new(report_to: person_to, report_from: person_from)
-      expect(report.valid?).to be_truthy
+  describe 'validations' do
+    it { is_expected.to belong_to(:report_to) }
+    it { is_expected.to belong_to(:report_from) }
+  end
+
+  context 'when the report is valid' do
+    it 'is a valid report' do
+      report = described_class.new(report_to: person_to, report_from: person_from)
+      expect(report).to be_valid
     end
   end
 
-  context "Invalid Report" do
-    it "It is not a valid report, person_to is nil" do
-      report = ReportPerson.new(report_to: nil, report_from: person_from)
+  context 'when the report is invalid' do
+    it 'is not a valid report, person_to is nil' do
+      report = described_class.new(report_to: nil, report_from: person_from)
 
-      expect(report.valid?).to be_falsy
+      expect(report).not_to be_valid
     end
 
-    it "It is not a valid report, person_from is nil" do
-      report = ReportPerson.new(report_to: person_to, report_from: nil)
+    it 'is not a valid report, person_from is nil' do
+      report = described_class.new(report_to: person_to, report_from: nil)
 
-      expect(report.valid?).to be_falsy
+      expect(report).not_to be_valid
     end
 
-    it "It is not a valid report, person to and from are the same" do
-      report = ReportPerson.new(report_to: person_from, report_from: person_from)
-      expect(report.valid?).to be_falsy
+    it 'is not a valid report, person to and from are the same' do
+      report = described_class.new(report_to: person_from, report_from: person_from)
+      expect(report).not_to be_valid
     end
 
-    it "It is not a valid report, already exists" do
-      ReportPerson.new(report_to: person_to, report_from: person_from).save
+    it 'is not a valid report, already exists' do
+      described_class.new(report_to: person_to, report_from: person_from).save
 
-      report = ReportPerson.new(report_to: person_to, report_from: person_from)
-      expect(report.valid?).to be_falsy
+      report = described_class.new(report_to: person_to, report_from: person_from)
+      expect(report).not_to be_valid
     end
   end
 end
