@@ -1,11 +1,11 @@
-require "rails_helper"
+require 'rails_helper'
 
-RSpec.describe "Reports", type: :request do
-  describe "GET /api/reports" do
-    context "all the reports" do
-      it "get reports with status code 200" do
-        get "/api/reports"
-        expect(response).to have_http_status(200)
+RSpec.describe 'Reports' do
+  describe 'GET /api/reports' do
+    context 'when fetching all the reports' do
+      it 'get reports with status code 200' do
+        get '/api/reports'
+        expect(response).to have_http_status(:ok)
       end
     end
   end
