@@ -1,12 +1,12 @@
-require "rails_helper"
+require 'rails_helper'
 
-RSpec.describe "Trade Items", type: :request do
-  describe "post /api/trade_items" do
-    context "valid attributes" do
+RSpec.describe 'Trade Items' do
+  describe 'post /api/trade_items' do
+    context 'with valid attributes' do
       let(:person_to)   { create(:person_with_inventory) }
       let(:person_from) { create(:person_with_inventory) }
 
-      it "perform a new trade sucessfully" do
+      it 'perform a new trade sucessfully' do
         trade_params_valid = {
           inventory_from:
           {
@@ -20,17 +20,17 @@ RSpec.describe "Trade Items", type: :request do
           }
         }
 
-        post "/api/trade_items", params: trade_params_valid
-        expect(response).to have_http_status(200)
+        post '/api/trade_items', params: trade_params_valid
+        expect(response).to have_http_status(:ok)
       end
     end
   end
 
-  context "invalid attributes" do
+  context 'with invalid attributes' do
     let(:person_to)   { create(:person_with_inventory) }
     let(:person_from) { create(:person_infected_with_inventory) }
-    
-    it "failed to perform a new trade, person from is infected" do
+
+    it 'failed to perform a new trade, person from is infected' do
       trade_params_valid = {
         inventory_from:
         {
@@ -44,8 +44,8 @@ RSpec.describe "Trade Items", type: :request do
         }
       }
 
-      post "/api/trade_items", params: trade_params_valid
-      expect(response).to have_http_status(422)
+      post '/api/trade_items', params: trade_params_valid
+      expect(response).to have_http_status(:unprocessable_content)
     end
   end
 end
